+++ packages/stripe_web/lib/src/web_stripe.dart
@@ -1,7 +1,6 @@
 //@dart=2.12
 import 'dart:async';
 import 'dart:developer' as dev;
-import 'package:web/web.dart' as web;
 import 'dart:ui' as ui;
 
 import 'package:flutter/widgets.dart';
@@ -10,6 +9,7 @@ import 'package:flutter_stripe_web/platform_pay_button.dart';
 import 'package:flutter_web_plugins/flutter_web_plugins.dart';
 import 'package:stripe_js/stripe_api.dart' as stripe_js;
 import 'package:stripe_js/stripe_js.dart' as stripe_js;
+import 'package:web/web.dart' as web;
 
 import 'parser/payment_intent.dart';
 import 'parser/payment_methods.dart';
@@ -271,7 +271,18 @@ class WebStripe extends StripePlatform {
   Future<PaymentIntent> handleNextAction(String paymentIntentClientSecret,
       {String? returnURL}) async {
     final stripe_js.PaymentIntentResponse response =
+        await _stripe.handleNextAction(paymentIntentClientSecret);
+
+    return response.paymentIntent!.parse();
+  }
+
+  @override
+  Future<PaymentIntent> handleCardAction(
+    String paymentIntentClientSecret,
+  ) async {
+    final stripe_js.PaymentIntentResponse response =
         await _stripe.handleCardAction(paymentIntentClientSecret);
+
     return response.paymentIntent!.parse();
   }
 
@@ -349,8 +360,11 @@ class WebStripe extends StripePlatform {
   }
 
   @override
-  Future<PaymentIntent> retrievePaymentIntent(String clientSecret) async {
-    throw UnimplementedError();
+  Future<PaymentIntent> retrievePaymentIntent(final String clientSecret) async {
+    final stripe_js.PaymentIntentResponse response =
+        await _stripe.retrievePaymentIntent(clientSecret);
+
+    return response.paymentIntent!.parse();
   }
 
   @override
@@ -393,11 +407,13 @@ class WebStripe extends StripePlatform {
   }
 
   Future<PaymentIntent> confirmPaymentElement(
-    ConfirmPaymentElementOptions options,
-  ) async {
+    ConfirmPaymentElementOptions options, [
+    String? clientSecret,
+  ]) async {
     final response = await js.confirmPayment(
       stripe_js.ConfirmPaymentOptions(
         elements: elements!,
+        clientSecret: clientSecret,
         confirmParams: options.confirmParams,
         redirect: options.redirect,
       ),
@@ -406,6 +422,26 @@ class WebStripe extends StripePlatform {
       throw response.error!;
     } else {
       return response.paymentIntent!.parse();
+    }
+  }
+
+  Future<void> elementsSubmit() => elements!.submit();
+
+  @override
+  Future<PaymentMethod> createPaymentMethodWithElements() async {
+    try {
+      final response = await js.createPaymentMethodWithElements(
+        stripe_js.CreatePaymentMethodWithElementsData(elements: elements!),
+      );
+
+      if (response.error != null) {
+        throw response.error!;
+      }
+
+      return response.paymentMethod!.parse();
+    } catch (e) {
+      dev.log('Error $e');
+      rethrow;
     }
   }
 
@@ -502,6 +538,24 @@ class WebStripe extends StripePlatform {
   @override
   Future<void> resetPaymentSheetCustomer() {
     throw WebUnsupportedError.method('resetPaymentSheet');
+  }
+
+  @override
+  Future<AvailableMobilePayOptions> availableMobilePayOptions({
+    IsGooglePaySupportedParams? params,
+    PlatformPayWebPaymentRequestCreateOptions? paymentRequestOptions,
+  }) async {
+    final paymentRequest = js.paymentRequest((paymentRequestOptions ??
+            PlatformPayWebPaymentRequestCreateOptions.defaultOptions)
+        .toJS());
+
+    final paymentOptions = await paymentRequest.canMakePayment();
+
+    return AvailableMobilePayOptions(
+      googlePay: paymentOptions?.googlePay ?? false,
+      applePay: paymentOptions?.applePay ?? false,
+      link: paymentOptions?.link ?? false,
+    );
   }
 
   @override