+++ packages/stripe_web/lib/src/widgets/payment_element.dart
@@ -9,13 +9,27 @@ import 'package:web/web.dart' as web;
 
 import '../../flutter_stripe_web.dart';
 
-export 'package:stripe_js/stripe_api.dart' show PaymentElementLayout;
+export 'package:stripe_js/stripe_api.dart'
+    show
+        PaymentElementLayout,
+        PaymentElementDefaultValues,
+        PaymentElementBillingDetails,
+        PaymentElementBillingDetailsAddress;
 
 typedef PaymentElementTheme = js.ElementTheme;
+typedef PaymentElementAppearance = js.ElementAppearance;
+typedef PaymentElementAppearanceLabels = js.ElementAppearanceLabels;
 
 class PaymentElement extends StatefulWidget {
-  final String clientSecret;
+  final String? clientSecret;
   final String? customerSessionClientSecret;
+
+  final int? amount;
+  final String? currency;
+  final String? mode;
+  final String? paymentMethodCreation;
+  final List<String>? paymentMethodTypes;
+
   final double? width;
   final double? height;
   final CardStyle? style;
@@ -35,11 +49,17 @@ class PaymentElement extends StatefulWidget {
   final js.PaymentElementOptionsTerms? terms;
   final js.PaymentElementWalletOptions? wallets;
   final js.PaymentElementApplePayOptions? applePay;
+  final String? locale;
 
   PaymentElement({
     super.key,
-    required this.clientSecret,
+    this.clientSecret,
     this.customerSessionClientSecret,
+    this.amount,
+    this.currency,
+    this.mode,
+    this.paymentMethodTypes,
+    this.paymentMethodCreation,
     this.width,
     this.height,
     this.style,
@@ -51,6 +71,7 @@ class PaymentElement extends StatefulWidget {
     required this.onCardChanged,
     this.layout = PaymentElementLayout.accordion,
     this.appearance,
+    this.locale,
     this.defaultValues,
     this.business,
     this.paymentMethodOrder,
@@ -121,7 +142,7 @@ class PaymentElementState extends State<PaymentElement> {
       ..style.border = 'none'
       ..style.width = '100%'
       ..style.height = '${height}'
-      ..style.overflow = 'scroll'
+      ..style.overflow = 'auto'
       ..style.overflowX = 'hidden';
 
     elements = WebStripe.js.elements(createOptions());
@@ -185,10 +206,27 @@ class PaymentElementState extends State<PaymentElement> {
 
   js.JsElementsCreateOptions createOptions() {
     final appearance = widget.appearance ?? js.ElementAppearance();
+
+    if (widget.clientSecret?.isNotEmpty == true) {
+      return js.JsElementsCreateOptions(
+        clientSecret: widget.clientSecret,
+        customerSessionClientSecret: widget.customerSessionClientSecret,
+        appearance: appearance.toJson().jsify() as js.JsElementAppearance,
+        locale: widget.locale,
+      );
+    }
+
     return js.JsElementsCreateOptions(
-      clientSecret: widget.clientSecret,
-      customerSessionClientSecret: widget.customerSessionClientSecret,
+      amount: widget.amount,
+      currency: widget.currency,
+      mode: widget.mode,
+      paymentMethodTypes: widget.paymentMethodTypes
+          ?.map((pmt) => pmt.toJS)
+          .toList(growable: false)
+          .toJS,
+      paymentMethodCreation: widget.paymentMethodCreation,
       appearance: appearance.toJson().jsify() as js.JsElementAppearance,
+      locale: widget.locale,
     );
   }