+++ packages/stripe/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 11.1.1
+
+**Features**
+- Add deferred payment methods for web
+
 ## 11.1.0
 - Sync with Stripe [0.38.6](https://github.com/stripe/stripe-react-native/releases/tag/v0.38.6). 
 - Minor fixes and improvements.